+++ tasks/common.py
@@ -1,4 +1,8 @@
 # -*- coding: utf-8 -*-
+# This file has been generated with 'invoke project.sync'.
+# Do not modify. Any manual change will be lost.
+# Please propose your modification on
+# https://github.com/camptocamp/odoo-template instead.
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
@@ -50,8 +54,8 @@ TEMPLATE_GIT = TEMPLATE_GIT_REPO_URL.format('camptocamp/odoo-template')
 
 def gpg_decrypt_to_file(ctx, file_name, password=False):
     """Get a value from lastpass.
-    :param password: password to decript gpg file
     :param file_name: File .gpg to decrypt
+    :param password: password to decript gpg file
     """
     passphrase = ""
     if password: