+++ tasks/pr.py
@@ -1,3 +1,7 @@
+# This file has been generated with 'invoke project.sync'.
+# Do not modify. Any manual change will be lost.
+# Please propose your modification on
+# https://github.com/camptocamp/odoo-template instead.
 # Copyright 2020 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
@@ -5,7 +9,7 @@ import os
 
 from invoke import exceptions, task
 
-from .common import check_git_diff, exit_msg
+from .common import ask_confirmation, check_git_diff, exit_msg
 from .database import _get_list_of_dumps, download_dump
 from .submodule import init, update
 
@@ -14,33 +18,46 @@ from .submodule import init, update
 def test(
     ctx,
     pr_link,
-    get_production_db=None,
-    get_integration_db=None,
+    get_local_db=None,
+    get_remote_db=None,
     template_db=None,
     create_template=None,
     keep_alive=None,
     base_branch='master',
+    # legacy args to show a nice error message
+    get_production_db=None,
+    get_integration_db=None,
 ):
     """
     This task will start a project build with requested PR and database
     datebase can be loaded from production/integration or reused previous
     created database template
     """
-    _check_arguments(
-        get_production_db, get_integration_db, template_db, create_template
-    )
+    _check_arguments(get_local_db, get_remote_db, template_db, create_template)
     docker_yml_name = 'docker-compose.override-{}.yml'.format(pr_link)
     if not keep_alive:
         restart(ctx)
     handle_git_repository(ctx, pr_link, base_branch)
     generate_docker_yml(pr_link, docker_yml_name)
 
-    database = get_integration_db or get_integration_db
-    if database:
-        database_dumb = _load_database(ctx, pr_link, database)
+    if get_remote_db:
+        dump_path = _get_list_of_dumps(ctx, get_remote_db)[-1]
+        gpg_fname = os.path.basename(dump_path)
+        fname = os.path.splitext(gpg_fname)[0]
 
-    if create_template:
-        _handle_database_template(ctx, pr_link, database_dumb)
+        download_dump(ctx, get_remote_db)
+    elif get_local_db:
+        fname = get_local_db
+    else:
+        if ask_confirmation(
+            'No --get-local-db/--get-remote-db parameter specified - build database from scratch?'
+        ):
+            _create_db(ctx, 'odoodb-{}'.format(pr_link))
+    if fname:
+        database_dump = _load_database(ctx, pr_link, fname)
+
+    if create_template and database_dump:
+        _handle_database_template(ctx, pr_link, database_dump)
 
     if template_db:
         _restore_database_from_template(ctx, pr_link, template_db)
@@ -48,28 +65,51 @@ def test(
     print('Starting container')
     print('Database migration started you can reach database on port 8069')
     ctx.run(
-        'docker-compose -f docker-compose.yml -f {} up'.format(
-            docker_yml_name
-        ),
-        hide=True,
+        'docker-compose -f docker-compose.yml -f {} up'.format(docker_yml_name)
     )
 
 
+@task
+def clean(ctx, pr_number):
+    """
+    This task will clean the branch and database created for pr.test task
+    """
+    print('Removing branch')
+    try:
+        ctx.run('git checkout master')
+        ctx.run('git branch -D {}'.format(pr_number), hide=False)
+    except Exception as expt:
+        print("Error when trying to remove branch : {}".format(expt))
+
+    print('Removing database')
+    _drop_db(ctx, 'odoodb-{}'.format(pr_number))
+    _drop_db(ctx, 'odoodb{}-template'.format(pr_number))
+
+
 def _check_arguments(
-    get_production_db=None,
-    get_integration_db=None,
+    get_local_db=None,
+    get_remote_db=None,
     template_db=None,
     create_template=None,
+    # legacy args to show a nice error message
+    get_production_db=None,
+    get_integration_db=None,
 ):
     # check if all arguments correctly send
-    if get_production_db and get_integration_db:
+    if get_local_db and get_remote_db:
         print(
-            'You cannot initialize production and integration in the same time'
+            'You cannot initialize a database from two sources at the same time.'
         )
         raise exceptions.Exit(1)
 
-    if create_template and not (get_production_db or get_integration_db):
+    if create_template and not (get_local_db or get_remote_db):
         print('For template option database argument is required')
+        raise exceptions.Exit(1)
+
+    if get_production_db or get_integration_db:
+        print(
+            'get_production_db and get_integration_db are now deprecated. Please use get_local_db or get_remote_db.'
+        )
         raise exceptions.Exit(1)
 
 
@@ -82,7 +122,7 @@ def _restore_database_from_template(ctx, pr_link, template):
     )
 
 
-def _handle_database_template(ctx, pr_link, database_dumb):
+def _handle_database_template(ctx, pr_link, database_dump):
     # at this point we should have the database loaded under proper name
     _drop_db(ctx, 'odoodb{}-template'.format(pr_link))
     _create_db(ctx, 'odoodb{}-template'.format(pr_link))
@@ -91,7 +131,7 @@ def _handle_database_template(ctx, pr_link, database_dumb):
         print('Creating template')
         ctx.run(
             'docker-compose run --rm odoo pg_restore -p 5432 -d odoodb{}-template < {}'.format(
-                pr_link, database_dumb
+                pr_link, database_dump
             ),
             hide=True,
         )
@@ -100,13 +140,7 @@ def _handle_database_template(ctx, pr_link, database_dumb):
         pass
 
 
-def _load_database(ctx, pr_link, database):
-
-    dump_path = _get_list_of_dumps(ctx, database)[-1]
-    gpg_fname = os.path.basename(dump_path)
-    fname = os.path.splitext(gpg_fname)[0]
-
-    download_dump(ctx, database)
+def _load_database(ctx, pr_link, fname):
     _drop_db(ctx, 'odoodb-{}'.format(pr_link))
     _create_db(ctx, 'odoodb-{}'.format(pr_link))
 
@@ -120,7 +154,6 @@ def _load_database(ctx, pr_link, database):
                 hide=True,
             )
         except Exception:
-            # to ignore warnings on db restore
             pass
     else:
         msg = "** Database file for restore is not found**"
@@ -187,6 +220,7 @@ services:
   odoo:
     environment:
         DB_NAME: odoodb-{}
+        MARABUNTA_MODE: full
   nginx:
     ports:
       - 8069:80